+++ tests/regression_tests_data.py
@@ -13,7 +13,7 @@
 # File contributors : Étienne André
 #
 # Created           : 2015/10/23
-# Last modified     : 2019/05/30
+# Last modified     : 2019/07/09
 #************************************************************